new Commtt
May have slightly butchered it, but almost done my portion. if antybody wants to finish please ask me questions. I have added code and methods to the Project Model class to get process (from list or anywhere else) and to feed processes to the model (right now just from a file but will add more ways. as well as some utility functinos to it. and have added code to BuddySystemStrategy class, new instaance of Memory Block which can be restructued. and have started the AllocateProcess function and will finish (if nobody else wants to) when i get a chance.