new fitting pettern *READ DESCRIPTION*
So I added the Alloc Strategy that william sent in java. I am finishing adding the the allocate method...
 If anybody else wants to take a stab at it.
Also, The deallocate method should be the same as the buddy system, because we are not using a specific algorithm to deallocate space, we are just deallocating it.

Please let me know if there are any questions.

- chris